fixed minor inconsistancis on variable names and type casting

+++ prepare_data/gen_24net_data.py
@@ -54,8 +54,8 @@ for annotation in annotations:
             delta_x = npr.randint(-w * 0.2, w * 0.2)
             delta_y = npr.randint(-h * 0.2, h * 0.2)
 
-            nx1 = max(x1 + w / 2 + delta_x - size / 2, 0)
-            ny1 = max(y1 + h / 2 + delta_y - size / 2, 0)
+            nx1 = max(int(x1 + w / 2 + delta_x - size / 2), 0)
+            ny1 = max(int(y1 + h / 2 + delta_y - size / 2), 0)
             nx2 = nx1 + size
             ny2 = ny1 + size
